+++ app/src/core/store.ts
@@ -1,7 +1,7 @@
 import { reactive, toRaw } from "vue";
 import type { FolderObject } from "@estoc/folder-object";
 import type { Imported } from "@estoc/event-store";
-import type { Delivery } from "@estoc/vault/v2";
+import type { Delivery } from "@estoc/vault";
 import {
   BASIC_MESSAGE,
   invitationMessage,